+++ HISTORY.rst
@@ -7,6 +7,7 @@ unreleased
 ----------
 
 - Add user's favorite albums, artists, tracks and playlists in user resource.
+- Add ``headers`` kwarg in Client to force session headers.
 
 1.1.2 (2019-06-02)
 ------------------